Updated build-script

+++ d/drupal-eu_cookie_compliance/drupal-eu_cookie_compliance_ubi_8.5.sh
@@ -19,7 +19,9 @@
 # ----------------------------------------------------------------------------
 # Variables
 PACKAGE_NAME=eu_cookie_compliance
-PACKAGE_URL=https://git.drupalcode.org/project/eu-cookie-compliance/
+CORE_PACKAGE_NAME=drupal
+PACKAGE_URL=https://git.drupalcode.org/project/eu-cookie-compliance
+CORE_PACKAGE_URL=https://github.com/drupal/drupal
 #PACKAGE_VERSION is configurable can be passed as an argument.
 PACKAGE_VERSION=${1:-8.x-1.14}
 
@@ -30,11 +32,39 @@ php -r "copy('https://getcomposer.org/installer', 'composer-setup.php');" && php
 OS_NAME=`cat /etc/os-release | grep PRETTY_NAME | cut -d '=' -f2 | tr -d '"'`
 
 #Check if package exists
+if [ -d "$CORE_PACKAGE_NAME" ] ; then
+  rm -rf $CORE_PACKAGE_NAME
+  echo "$CORE_PACKAGE_NAME  | $PACKAGE_VERSION | $OS_NAME | GitHub | Removed existing package if any"
+
+fi
+
+if ! git clone $CORE_PACKAGE_URL $CORE_PACKAGE_NAME; then
+        echo "------------------$PACKAGE_NAME:clone_fails---------------------------------------"
+                echo "$CORE_PACKAGE_URL $CORE_PACKAGE_NAME"
+        echo "$CORE_PACKAGE_NAME  |  $CORE_PACKAGE_URL |  $PACKAGE_VERSION | $OS_NAME | GitHub | Fail |  Clone_Fails"
+        exit 1
+fi
+cd $CORE_PACKAGE_NAME
+git checkout 8.9.0
+rm -rf composer.lock
+
+composer config allow-plugins true --no-interaction
+
+if ! composer install --no-interaction; then
+        echo "------------------$PACKAGE_NAME:install_fails-------------------------------------"
+        echo "$PACKAGE_URL $PACKAGE_NAME"
+        echo "$PACKAGE_NAME  |  $PACKAGE_URL | $PACKAGE_VERSION | $OS_NAME | GitHub | Fail |  Install_Fails"
+        exit 1
+fi
+    composer require --dev phpunit/phpunit --with-all-dependencies ^7 --no-interaction
+    cd modules/
+
+#Check if package exists
 if [ -d "$PACKAGE_NAME" ] ; then
   rm -rf $PACKAGE_NAME
   echo "$PACKAGE_NAME  | $PACKAGE_VERSION | $OS_NAME | GitHub | Removed existing package if any"
-
 fi
+
 if ! git clone $PACKAGE_URL $PACKAGE_NAME; then
         echo "------------------$PACKAGE_NAME:clone_fails---------------------------------------"
                 echo "$PACKAGE_URL $PACKAGE_NAME"
@@ -43,13 +73,18 @@ if ! git clone $PACKAGE_URL $PACKAGE_NAME; then
 fi
 cd $PACKAGE_NAME
 git checkout $PACKAGE_VERSION
-composer require --dev phpunit/phpunit --with-all-dependencies ^7 --no-interaction
+cd ../../
+cd core/
 
-if ! composer install --no-interaction; then
-        echo "------------------$PACKAGE_NAME:install_fails-------------------------------------"
+if !  ../vendor/phpunit/phpunit/phpunit ../modules/$PACKAGE_NAME/; then
+        echo "------------------$PACKAGE_NAME:install_success_but_test_fails---------------------"
         echo "$PACKAGE_URL $PACKAGE_NAME"
-        echo "$PACKAGE_NAME  |  $PACKAGE_URL | $PACKAGE_VERSION | $OS_NAME | GitHub | Fail |  Install_Fails"
+        echo "$PACKAGE_NAME  |  $PACKAGE_URL | $PACKAGE_VERSION | $OS_NAME | GitHub | Fail |  Install_success_but_test_Fails"
         exit 1
+else
+        echo "------------------$PACKAGE_NAME:install_&_test_both_success-------------------------"
+        echo "$PACKAGE_URL $PACKAGE_NAME"
+        echo "$PACKAGE_NAME  |  $PACKAGE_URL | $PACKAGE_VERSION | $OS_NAME | GitHub  | Pass |  Both_Install_and_Test_Success"
+        exit 0
 fi
-
 # No test case available.
